Fix tarball build
Another casualty of the enabling of pipefail in stdenv.

http://hydra.nixos.org/build/16918723

+++ release.nix
@@ -37,7 +37,9 @@ rec {
 
     postUnpack = ''
       # Clean up when building from a working tree.
-      (cd $sourceRoot && (git ls-files -o | xargs -r rm -v))
+      if [ -d $sourceRoot/.git ]; then
+        (cd $sourceRoot && (git ls-files -o | xargs -r rm -v))
+      fi
     '';
 
     distPhase =
